requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
asgiref==3.3.1; python_version >= '3.5'
django-crispy-forms==1.11.2
django-email-verification==0.1.0
django-filter==2.4.0
django==3.1.7
pillow==8.1.2
pytz==2021.1
sqlparse==0.4.1; python_version >= '3.5'
